+++ test/OpenTelemetry.Resources.ProcessRuntime.Tests/ProcessRuntimeDetectorTests.cs
@@ -2,10 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System.Linq;
-using OpenTelemetry.Resources;
 using Xunit;
 
-namespace OpenTelemetry.ResourceDetectors.ProcessRuntime.Tests;
+namespace OpenTelemetry.Resources.ProcessRuntime.Tests;
 
 public class ProcessRuntimeDetectorTests
 {