+++ src/OpenTelemetry.Resources.ProcessRuntime/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* **Breaking Change**: Renamed package from `OpenTelemetry.ResourceDetectors.ProcessRuntime`
+  to `OpenTelemetry.Resources.ProcessRuntime`.
+  ([#1767](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1767))
+* **Breaking Change**: `ProcessRuntimeDetector` type is now internal, use `ResourceBuilder`
+  extension method `AddProcessRuntimeDetector` to enable the detector.
+  ([#1767](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1767))
 * Update OpenTelemetry SDK version to `1.8.1`.
   ([#1668](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1668))